Wire vesting percent

+++ src/services/src/Demoulas.ProfitSharing.Common/ReferenceData.cs
@@ -2,6 +2,6 @@
 public static class ReferenceData
 {
     // Vesting Schedule Percent by year
-    public static readonly List<int> OlderVestingSchedule = [0, 0, 20, 40, 60, 80, 100];
-    public static readonly List<int> NewerVestingSchedule = [0, 20, 40, 60, 80, 100];
+    public static readonly List<byte> OlderVestingSchedule = [0, 0, 20, 40, 60, 80, 100];
+    public static readonly List<byte> NewerVestingSchedule = [0, 20, 40, 60, 80, 100];
 }
